Polish: migrate remaining legacy colors, extract test fixture
- SettingsScreen: clear-orange/clear-lime → semantic tokens
- OnboardingScreen: clear-orange → --icon-cta and --text-card-label
- SessionDetailScreen: --border-card → --text-card-label, raw green → --text-timer
- Extract generateTestWorkout from workout.ts to lib/test-fixtures.ts
- Remove unused sectionStartedAt and totalSections props
- ExerciseNotes: fix --text-color → --text-paragraph, add aria-label
- auth-storage: capture getStayLoggedIn() once per method
