+++ src/lib/test-fixtures.ts
@@ -1,0 +1,304 @@
+import { GeneratedWorkout } from "@/types/workout";
+
+/**
+ * Test workout covering every structure type.
+ * Used only by TestWorkoutScreen for development.
+ */
+export const generateTestWorkout = (): GeneratedWorkout => {
+  const circuitId = crypto.randomUUID();
+  return {
+    id: crypto.randomUUID(),
+    title: "Test Mega Workout",
+    description: "Covers every structure type: standard, superset, circuit, EMOM, AMRAP, for_time, and ladder reps.",
+    duration: "60m",
+    intensity: 7,
+    anchor: "full body",
+    sections: [
+      // 1. Warm-Up — standard
+      {
+        id: "sec-warmup",
+        name: "Warm-Up",
+        type: "warmup",
+        status: "not_started",
+        exercises: [
+          {
+            id: "wu-1",
+            name: "Jumping Jacks",
+            sets: 2,
+            reps: "20",
+            rest: "15s",
+            coachingCues: ["Light bounce, arms fully overhead"],
+            structure: { type: "standard" },
+          },
+          {
+            id: "wu-2",
+            name: "Inchworm Walk-Out",
+            sets: 2,
+            reps: "6",
+            tempo: "2-1-2",
+            rest: "20s",
+            coachingCues: ["Walk hands out to plank", "Squeeze glutes at bottom"],
+            regression: "Walk out to half plank",
+            structure: { type: "standard" },
+          },
+        ],
+      },
+      // 2. Mobility Flow — standard with full metadata
+      {
+        id: "sec-mobility",
+        name: "Mobility Flow",
+        type: "mobility",
+        status: "not_started",
+        exercises: [
+          {
+            id: "mob-1",
+            name: "World's Greatest Stretch",
+            sets: 2,
+            reps: "5 each",
+            tempo: "3-2-3",
+            effort: "RPE 3",
+            rest: "0s",
+            equipment: "Mat",
+            coachingCues: ["Deep lunge, rotate toward front knee", "Open chest to ceiling"],
+            regression: "Half-kneeling version",
+            progression: "Add overhead reach",
+            structure: { type: "standard" },
+          },
+          {
+            id: "mob-2",
+            name: "90/90 Hip Switch",
+            sets: 2,
+            reps: "8 each",
+            tempo: "2-1-2",
+            rest: "15s",
+            equipment: "Mat",
+            coachingCues: ["Sit tall, rotate from hips"],
+            structure: { type: "standard" },
+          },
+        ],
+      },
+      // 3. Primary Superset — superset (two exercises paired)
+      {
+        id: "sec-primary",
+        name: "Primary Superset",
+        type: "primary",
+        status: "not_started",
+        exercises: [
+          {
+            id: "pri-1",
+            name: "Barbell Back Squat",
+            sets: 4,
+            reps: "6",
+            effort: "75%",
+            tempo: "3-1-1",
+            rest: "30s",
+            lastWeight: "185lbs",
+            equipment: "Barbell, Squat Rack",
+            coachingCues: ["Brace core, knees track toes", "Depth to parallel or below"],
+            regression: "Goblet Squat",
+            progression: "Pause Squat (3s hole)",
+            structure: { type: "superset", paired_with: "pri-2" },
+          },
+          {
+            id: "pri-2",
+            name: "Barbell Romanian Deadlift",
+            sets: 4,
+            reps: "8",
+            effort: "65%",
+            tempo: "3-1-1",
+            rest: "90s",
+            lastWeight: "155lbs",
+            equipment: "Barbell",
+            coachingCues: ["Hinge at hips, soft knees", "Bar stays close to legs"],
+            regression: "Dumbbell RDL",
+            progression: "Deficit RDL",
+            structure: { type: "superset", paired_with: "pri-1" },
+          },
+        ],
+      },
+      // 4. Accessory Circuit — circuit (3 exercises, 3 rounds)
+      {
+        id: "sec-accessory",
+        name: "Accessory Circuit",
+        type: "accessory",
+        status: "not_started",
+        exercises: [
+          {
+            id: "acc-1",
+            name: "Dumbbell Lateral Raise",
+            sets: null,
+            reps: "12",
+            effort: "RPE 7",
+            rest: "0s",
+            lastWeight: "15lbs",
+            equipment: "Dumbbells",
+            coachingCues: ["Slight lean forward, lead with pinkies"],
+            structure: { type: "circuit", circuit_id: circuitId, rounds: 3 },
+          },
+          {
+            id: "acc-2",
+            name: "Cable Face Pull",
+            sets: null,
+            reps: "15",
+            effort: "RPE 7",
+            rest: "0s",
+            equipment: "Cable Machine",
+            coachingCues: ["Pull to forehead, external rotate at end"],
+            structure: { type: "circuit", circuit_id: circuitId, rounds: 3 },
+          },
+          {
+            id: "acc-3",
+            name: "Hammer Curl",
+            sets: null,
+            reps: "10 each",
+            effort: "RPE 8",
+            rest: "60s",
+            lastWeight: "25lbs",
+            equipment: "Dumbbells",
+            coachingCues: ["Neutral grip, controlled eccentric"],
+            structure: { type: "circuit", circuit_id: circuitId, rounds: 3 },
+          },
+        ],
+      },
+      // 5. Skill / Power — standard with equipment & lastWeight
+      {
+        id: "sec-skill",
+        name: "Skill / Power",
+        type: "skill",
+        status: "not_started",
+        exercises: [
+          {
+            id: "sk-1",
+            name: "Kettlebell Swing",
+            sets: 4,
+            reps: "12",
+            effort: "RPE 8",
+            rest: "60s",
+            lastWeight: "53lbs",
+            equipment: "Kettlebells",
+            coachingCues: ["Hip snap, not a squat", "Arms are hooks, power from glutes", "Bell to chest height"],
+            regression: "Russian Swing (eye level)",
+            progression: "Single-Arm Swing",
+            structure: { type: "standard" },
+          },
+        ],
+      },
+      // 6. Core EMOM — emom (6 min)
+      {
+        id: "sec-core",
+        name: "Core EMOM",
+        type: "core",
+        status: "not_started",
+        exercises: [
+          {
+            id: "core-1",
+            name: "Hollow Body Hold",
+            sets: null,
+            reps: "30s",
+            coachingCues: ["Low back pressed to floor", "Arms by ears"],
+            structure: { type: "emom", minutes: 6 },
+          },
+          {
+            id: "core-2",
+            name: "V-Up",
+            sets: null,
+            reps: "10",
+            coachingCues: ["Touch toes at top, control down"],
+            structure: { type: "emom", minutes: 6 },
+          },
+        ],
+      },
+      // 7. AMRAP Conditioning — amrap (8 min)
+      {
+        id: "sec-conditioning",
+        name: "AMRAP Conditioning",
+        type: "conditioning",
+        status: "not_started",
+        exercises: [
+          {
+            id: "cond-1",
+            name: "Burpee",
+            sets: null,
+            reps: "8",
+            coachingCues: ["Chest to floor, jump at top"],
+            structure: { type: "amrap", minutes: 8 },
+          },
+          {
+            id: "cond-2",
+            name: "Box Jump",
+            sets: null,
+            reps: "10",
+            equipment: "Plyo Box",
+            coachingCues: ["Land soft, stand tall at top"],
+            regression: "Box Step-Up",
+            structure: { type: "amrap", minutes: 8 },
+          },
+          {
+            id: "cond-3",
+            name: "Kettlebell Goblet Squat",
+            sets: null,
+            reps: "12",
+            lastWeight: "35lbs",
+            equipment: "Kettlebells",
+            coachingCues: ["Elbows inside knees at bottom"],
+            structure: { type: "amrap", minutes: 8 },
+          },
+        ],
+      },
+      // 8. Stability Ladder For Time — for_time (5 min cap)
+      {
+        id: "sec-ladder-for-time",
+        name: "Ladder For Time",
+        type: "stability",
+        status: "not_started",
+        exercises: [
+          {
+            id: "stab-1",
+            name: "Single-Leg Romanian Deadlift",
+            sets: null,
+            reps: "2-4-6-8-6-4-2 each",
+            equipment: "Dumbbells",
+            coachingCues: ["Hinge from hip, reach toward floor", "Keep hips square"],
+            regression: "Kickstand RDL",
+            progression: "Eyes closed",
+            structure: { type: "for_time", time_cap_mins: 5 },
+          },
+          {
+            id: "stab-2",
+            name: "Pallof Press",
+            sets: null,
+            reps: "2-4-6-8-6-4-2 each",
+            equipment: "Cable Machine",
+            coachingCues: ["Press out slowly, resist rotation"],
+            structure: { type: "for_time", time_cap_mins: 5 },
+          },
+        ],
+      },
+      // 9. Cooldown — standard, minimal
+      {
+        id: "sec-cooldown",
+        name: "Cooldown",
+        type: "cooldown",
+        status: "not_started",
+        exercises: [
+          {
+            id: "cd-1",
+            name: "Child's Pose",
+            sets: 1,
+            reps: "60s",
+            coachingCues: ["Sink hips to heels, walk hands forward"],
+            structure: { type: "standard" },
+          },
+          {
+            id: "cd-2",
+            name: "Supine Spinal Twist",
+            sets: 1,
+            reps: "45s each",
+            coachingCues: ["Both shoulders on the floor"],
+            structure: { type: "standard" },
+          },
+        ],
+      },
+    ],
+  };
+};
